fix(null-renderer): promote w*h to size_t before multiplication
w * h * 3 and w * h were computed as signed int before the size_t cast,
causing undefined behaviour on large dimensions.  Cast each operand to
size_t individually so the multiplication is performed in size_t.

+++ renderer/null/NullRenderBackend.h
@@ -43,14 +43,14 @@ public:
     bool ReadbackColorBgr8(int w, int h, std::vector<uint8_t> &out,
                            std::string *) override {
         if (w <= 0 || h <= 0) return false;
-        out.assign(static_cast<size_t>(w * h * 3), 0u);
+        out.assign(static_cast<size_t>(w) * static_cast<size_t>(h) * 3u, 0u);
         return true;
     }
 
     bool ReadbackDepth32F(int w, int h, std::vector<float> &out,
                           std::string *) override {
         if (w <= 0 || h <= 0) return false;
-        out.assign(static_cast<size_t>(w * h), 1.0f);
+        out.assign(static_cast<size_t>(w) * static_cast<size_t>(h), 1.0f);
         return true;
     }
